+++ backend/matching-service/.eslintrc.json
@@ -3,6 +3,7 @@
     "es2021": true,
     "node": true
   },
+  "plugins": ["@typescript-eslint"],
   "extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
   "overrides": [],
   "parser": "@typescript-eslint/parser",
@@ -10,6 +11,5 @@
     "ecmaVersion": "latest",
     "sourceType": "module"
   },
-  "plugins": ["@typescript-eslint"],
   "rules": {}
 }
