+++ backend/matching-service/package.json
@@ -6,7 +6,8 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "npx tsc",
     "start": "node dist/index.js",
-    "dev": "nodemon -L src/index.ts"
+    "dev": "nodemon -L src/index.ts",
+    "lint": "eslint ."
   },
   "keywords": [],
   "author": "",
